requirements.txt
pandas==2.0.0
numpy==1.24.3
openai==1.3.0
python-dotenv==1.0.0
datetime==5.1
langchain==0.0.267
tabulate==0.9.0
matplotlib==3.7.2
seaborn==0.12.2 